EnigmaMind-django/requirements.txt
gpt_index==0.4.24
langchain==0.0.153
python-dotenv==1.0.0
django
django-cors-headers
gunicorn